flaskapp/requirements.txt
Flask==2.2.3
ibm-db==3.1.4
ibm-db-sa==0.4.0
python-dotenv==1.0.0
SQLAlchemy==2.0.10